+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexHeaderLineImpl.java
@@ -30,6 +30,7 @@ import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
 import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
+import com.intellij.openapi.util.TextRange;
 import java.util.Collection;
 
 public class ImpexHeaderLineImpl extends ASTWrapperPsiElement implements ImpexHeaderLine {
@@ -76,6 +77,12 @@ public class ImpexHeaderLineImpl extends ASTWrapperPsiElement implements ImpexHe
   @NotNull
   public Collection<ImpexValueLine> getValueLines() {
     return ImpexPsiUtil.getValueLines(this);
+  }
+
+  @Override
+  @NotNull
+  public TextRange getTableRange() {
+    return ImpexPsiUtil.getTableRange(this);
   }
 
 }