+++ src/com/intellij/idea/plugin/hybris/impex/psi/ImpexPsiUtil.kt
@@ -26,6 +26,7 @@ import com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils
 import com.intellij.idea.plugin.hybris.properties.PropertiesService
 import com.intellij.idea.plugin.hybris.system.type.psi.reference.result.*
 import com.intellij.openapi.project.DumbService
+import com.intellij.openapi.util.TextRange
 import com.intellij.openapi.util.text.StringUtil
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiPolyVariantReference
@@ -33,6 +34,8 @@ import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.childrenOfType
 import com.intellij.psi.util.parentOfType
 import com.intellij.psi.util.siblings
+import com.intellij.refactoring.suggested.endOffset
+import com.intellij.refactoring.suggested.startOffset
 import org.jetbrains.kotlin.utils.addToStdlib.indexOfOrNull
 
 fun getHeaderLine(element: ImpexFullHeaderParameter): ImpexHeaderLine? = PsiTreeUtil
@@ -74,8 +77,39 @@ fun getValueLines(element: ImpexHeaderLine): Collection<ImpexValueLine> {
     return valueLines
 }
 
+fun getTableRange(element: ImpexHeaderLine): TextRange {
+    val tableElements = ArrayDeque<PsiElement>()
+    var next = element.nextSibling
+
+    while (next != null) {
+        if (next is ImpexHeaderLine || next is ImpexUserRightsStart) {
+
+            // once all lines processed, we have to go back till last value line
+            var lastElement = tableElements.lastOrNull()
+            while (lastElement != null && lastElement !is ImpexValueLine) {
+                tableElements.removeLastOrNull()
+                lastElement = tableElements.lastOrNull()
+            }
+
+            next = null
+        } else {
+            // skip User Rights inside ImpEx statement
+            if (next !is ImpexUserRights) {
+                tableElements.add(next)
+            }
+            next = next.nextSibling
+        }
+    }
+
+    val endOffset = tableElements.lastOrNull()
+        ?.endOffset
+        ?: element.endOffset
+
+    return TextRange.create(element.startOffset, endOffset)
+}
+
 /**
- * This method will get value of the value group and if it's empty will check for value in default attribute
+ * This method will get value of the value group and if it's empty will check for value in the default attribute
  */
 fun computeValue(element: ImpexValueGroup): String? {
     val computedValue = element